requirements.txt
hydra-core
torch
torchvision
torchmetrics
